[#101] employ the flake8 PEP linting tool
* feat(ci): install setuptools via pip

Looks like something changed in the dependency set which was
causing CI to fail installing dependencies through poetry due
to not having a setuptools module available.

This patch to Circle CI's config installs setuptools directly
into the virtualenv we're using before installing via poetry,
which resolves the issue.

Signed-off-by: Kevin Morris <kevr@0cost.org>

* feat(poetry): add `black` dependency

Signed-off-by: Kevin Morris <kevr@0cost.org>

* feat(poetry): add `black` configuration

Limit line sizes to 79 with black. The `black` tool can be used to
format Python code in a unified fashion, and check source code
to see if it's formatted by black (which can be useful for CI
lint tests).

Signed-off-by: Kevin Morris <kevr@0cost.org>

* change: add 'home/static' to gitignore

Signed-off-by: Kevin Morris <kevr@0cost.org>

* change: format everything under ./home with black

Signed-off-by: Kevin Morris <kevr@0cost.org>

* change: format server/{settings,urls}.py with black

Signed-off-by: Kevin Morris <kevr@0cost.org>

* change: format manage.py with black

Signed-off-by: Kevin Morris <kevr@0cost.org>

* feat(ci): add lint check step, starting with black

Signed-off-by: Kevin Morris <kevr@0cost.org>

* feat(poetry): add 'flake8' dependency

flake8 is a linting tool used to detect violations of
PEP guidelines. It can be used to lint within CI and
by developers to avoid issues with code.

Signed-off-by: Kevin Morris <kevr@0cost.org>

* feat: add flake8 configuration file: .flake8

Signed-off-by: Kevin Morris <kevr@0cost.org>

* feat(poetry): add 'autoflake' dependency

A tool that can be used to fix/check flake8 violations.

Signed-off-by: Kevin Morris <kevr@0cost.org>

* change: format everything under ./home to pass flake8

Signed-off-by: Kevin Morris <kevr@0cost.org>

* change: fix 'manage.py' flake8 violations

Signed-off-by: Kevin Morris <kevr@0cost.org>

* feat(ci): add lint check step for flake8

Signed-off-by: Kevin Morris <kevr@0cost.org>

* Few more merge fixes

* More merge fixes

Signed-off-by: Kevin Morris <kevr@0cost.org>
Co-authored-by: Kevin Morris <kevr@0cost.org>