feat: POST /api/currency/get, a step currency backend
This route does a very simple thing: it converts the number of
steps that a device has into a coin + badge value, and presents
those currency values to the requesting user.

It will also provide us a "single source of truth" when it comes
to currency calculations for the application.

In the front-end, this route should be used to obtain a server-side,
up-to-date record of these gamified currencies. The value can then be
cached in client state to deal with user interaction animations.

Signed-off-by: Kevin Morris <kevr@0cost.org>