fix(users): default numeric columns for users with no walks to 0
This was a bit confusing: before this commit, the backend users page
displays new signups whether they've walked or not, but in cases
where they have not walked at all, no numeric value is displayed.

Anyway, TL;DR: This commit defaults numeric columns in user rows
with no walks to 0, which indicates that the user hasn't walked
and makes a bit more sense.

Signed-off-by: Kevin Morris <kevr@0cost.org>