Upgraded project to use latest django 4.2.x code (or higher, but not 5.anything).
Needed to update "django-postgres-setfield" as well to a custom
version that fixed a known issue with 'ugettext_lazy'. See:
https://code.djangoproject.com/ticket/30165.

Also regenerated the requirements.txt file per the above updates.

NOTES: In future, might want to update the Account model to have race
use models.JSONField() instead of this deprected SetField().